Move counter implementation from benchmark.cc to counter.cc.
    See discussion: https://github.com/google/benchmark/pull/262#discussion_r72298980 .
